requirements.txt
requests>=2.31.0
streamlit>=1.32.0
plotly>=5.19.0
pandas>=2.0.0
rich>=13.7.0
pytest>=8.0.0
python-dotenv>=1.0.0
